+++ TombEngine/Objects/TR5/Trap/LaserBeam.h
