+++ TombEngine/Objects/TR5/tr5_objects.cpp
@@ -57,7 +57,7 @@
 // Traps
 #include "Objects/Effects/tr5_electricity.h"
 #include "Objects/TR5/Trap/LaserBarrier.h"
-#include "Objects/TR5/Trap/SingleLaser.h"
+#include "Objects/TR5/Trap/LaserBeam.h"
 #include "Objects/TR5/Trap/ZipLine.h"
 #include "Objects/TR5/Object/tr5_rollingball.h"
 #include "Objects/TR5/Trap/tr5_ventilator.h"
